[#xx] v2: Define API service interfaces
Define interfaces that reflect service (Accounting, Container, Session)
declarations in the API.
